Add files via upload
This project is written in R Language on R Studio. It install and imports libraries like shiny, shinythemes, data.table, R Curl and Random Forest, import datasets and create a model and server that can be used to predict based on important variables like outlook, temperature, humidity and wind to predict when to play golf and when to not.